+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:apmg/presentation/home/menu_dashboard.dart';
 import 'package:apmg/presentation/pod/screen_pod.dart';
 import 'package:apmg/presentation/showcase/product_showcase.dart';
 import 'package:apmg/presentation/showcase/screen_showcase.dart';
-import 'package:apmg/presentation/invoice/invoice.dart';
+import 'package:apmg/screens/invoice.dart';
 import 'package:apmg/presentation/login/login_screen.dart';
 import 'package:apmg/presentation/login/splash_screen.dart';
 import 'package:apmg/core/theme.dart';
@@ -23,7 +23,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'APMG',
       theme: theme(),
-      home: Invoice(),
+      home: MenuDashboardPage(),
     );
   }
 }
